+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Failsafe JS cache storage: use a custom in-memory storage as fallback when access to `sessionStorage` is not allowed [#1742](https://github.com/opendatateam/udata/pull/1742)
 - Prevent errors when handling API errors without data/payload [#1743](https://github.com/opendatateam/udata/pull/1743)
 - Improve/fix validation error formatting on harvesting [#1745](https://github.com/opendatateam/udata/pull/1745)
+- Ensure daterange can be parsed from full iso datetime [#1748](https://github.com/opendatateam/udata/pull/1748)
 
 ## 1.4.0 (2018-06-06)